+++ src/game/gameObject.ts
@@ -1,6 +1,8 @@
-import { type Game } from "./game";
-import { type ObjectKind, type SurvivBitStream, TypeToId, type Orientation } from "../utils";
-import { type Body, type Vec2 } from "planck";
+import type { Game } from "./game";
+import { TypeToId } from "../utils/data";
+import type { ObjectKind, Orientation } from "../utils/constants";
+import type { SurvivBitStream } from "../utils/survivBitStream";
+import type { Body, Vec2 } from "planck";
 
 export abstract class GameObject {
     // For interop with subclasses
@@ -19,8 +21,8 @@ export abstract class GameObject {
 
     _position: Vec2;
     layer: number;
-    // why is this field optional?
-    orientation?: Orientation;
+
+    orientation: Orientation;
     scale = 1;
 
     dead = false;
@@ -44,7 +46,7 @@ export abstract class GameObject {
         if (this.typeString) this.typeId = TypeToId[typeString];
         this._position = position;
         this.layer = layer;
-        this.orientation = orientation;
+        this.orientation = orientation ?? 0;
         this.game = game;
     }
