+++ src/packets/receiving/joinPacket.ts
@@ -1,5 +1,5 @@
 import { ReceivingPacket } from "../receivingPacket";
-import { type SurvivBitStream } from "../../utils";
+import type { SurvivBitStream } from "../../utils/survivBitStream";
 
 export class JoinPacket extends ReceivingPacket {
     deserialize (stream: SurvivBitStream): void {
