+++ src/packets/receiving/spectatePacket.ts
@@ -1,5 +1,5 @@
 import { ReceivingPacket } from "../receivingPacket";
-import { type SurvivBitStream } from "../../utils";
+import type { SurvivBitStream } from "../../utils/survivBitStream";
 
 export class SpectatePacket extends ReceivingPacket {
     deserialize (stream: SurvivBitStream): void {
