+++ src/packets/receivingPacket.ts
@@ -1,4 +1,4 @@
-import { type SurvivBitStream } from "../utils";
+import type { SurvivBitStream } from "../utils/survivBitStream";
 import { type Player } from "../game/objects/player";
 
 export abstract class ReceivingPacket {
